Restructuring queries a little bit, filtering specifically for APPLICATION and SERVICE and removing logs.APM filter.
